fixed session state logic, it was broken before
